[refactor] restful 설계
SSR 프로젝트지만 비동기로 API를 호출을 했기 때문에 RESTful 네이밍으로 수정완료
