+++ eoullim-Project/src/main/resources/templates/booking/detail.mustache
@@ -128,7 +128,7 @@
 
              try {
                  // 3. 서버에 결제 준비 요청 (사전 검증용)
-                 const paymentKeyResponse = await fetch("/api/payment/prepare", {
+                 const paymentKeyResponse = await fetch("/api/user/payments/prepare", {
                      method: "POST",
                      headers: {
                          "Content-Type": "application/json"
@@ -161,7 +161,7 @@
                      m_redirect_url: window.location.href
                  }, async (response) => {
                      if (response.success) {
-                         const notified = await fetch("/api/payment/complete", {
+                         const notified = await fetch("/api/user/payments/complete", {
                              method: "POST",
                              headers: {
                                  "Content-Type": "application/json"
@@ -176,7 +176,7 @@
 
                          if (notified.ok) {
                              alert(`${resData.message}\n결제완료 페이지로 이동합니다.`);
-                             location.href  = `/booking/complete?code=${resData.bookingCode}`;
+                             location.href  = `/user/bookings/complete?code=${resData.bookingCode}`;
                          } else {
                              alert("결제 실패 다시시도해주세요.");
                              location.href  = `/place/${roomId}/room`
@@ -195,7 +195,7 @@
 
         async function paymentCancel(paymentKey) {
             try {
-                await fetch("/api/payment/cancel", {
+                await fetch("/api/user/payments/cancel", {
                    method: "POST",
                    headers: {
                        "Content-Type" : "application/json"
@@ -217,7 +217,7 @@
                 return;
             }
 
-            const response = await fetch(`/api/booking/cancel/${bookingCode}`, {
+            const response = await fetch(`/api/user/bookings/cancel/${bookingCode}`, {
                 method: "POST",
                 headers: {
                     "Content-Type" : "application/json"
